将手势控制无人机系统拆分为模块化架构，保持功能完整性，优化代码维护性
