+++ src/autonomus_drone_hand_gesture_project/gesture_config.json
@@ -18,11 +18,11 @@
   },
   "gesture": {
     "skin_lower_h": 0,
-    "skin_upper_h": 176,
+    "skin_upper_h": 170,
     "skin_lower_s": 0,
-    "skin_upper_s": 116,
+    "skin_upper_s": 70,
     "skin_lower_v": 0,
-    "skin_upper_v": 92,
+    "skin_upper_v": 75,
     "min_hand_area": 3000,
     "max_hand_area": 30000,
     "hand_ratio_threshold": 1.5,
